+++ ChangeLog.md
@@ -4,6 +4,7 @@ Starting with v1.31.6, this file will contain a record of major features and upd
 
 ## Upcoming
 - Add support for setting custom edge and node tooltips ([Link to PR](https://github.com/aws/graph-notebook/pull/227))
+- Fixed error causing query autocompletion to fail ([Link to PR](https://github.com/aws/graph-notebook/pull/231))
 
 ## Release 3.0.8 (November 3, 2021)
 - Added support for specifying the Gremlin traversal source ([Link to PR](https://github.com/aws/graph-notebook/pull/221))